Group Deadline and Event classes under abstract class TimedTask
